+++ BotFileCreator/BotFileCreationWizard.xaml.cs
@@ -1,9 +1,10 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-using System;
 
 namespace BotFileCreator
 {
+    using System;
+
     /// <summary>
     /// Interaction logic for BotFileCreationWizard.xaml
     /// </summary>
