+++ BotFileCreator/Data/EndpointItem.cs
@@ -1,4 +1,7 @@
-﻿namespace BotFileCreator
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace BotFileCreator
 {
     public class EndpointItem : BaseViewModel
     {
