+++ BotFileCreator/GeneralSettings.settings
@@ -1,10 +1,7 @@
 ﻿<?xml version='1.0' encoding='utf-8'?>
-<SettingsFile xmlns="http://schemas.microsoft.com/VisualStudio/2004/01/settings" CurrentProfile="(Default)" GeneratedClassNamespace="VSIXCustomPopUp" GeneratedClassName="GeneralSettings">
+<SettingsFile xmlns="http://schemas.microsoft.com/VisualStudio/2004/01/settings" CurrentProfile="(Default)" GeneratedClassNamespace="BotFileCreator" GeneratedClassName="GeneralSettings">
   <Profiles />
   <Settings>
-    <Setting Name="BotFileName" Type="System.String" Scope="User">
-      <Value Profile="(Default)" />
-    </Setting>
     <Setting Name="ProjectName" Type="System.String" Scope="User">
       <Value Profile="(Default)" />
     </Setting>
