--- BotFileCreator/MSBotCommands/MSBotCommandManager.cs
@@ -1,10 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-namespace BotFileCreator
-{
-    public abstract class MSBotCommandManager
-    {
-        public abstract string GetMSBotCommand();
-    }
-}
